Renaming `ElementConfigurationSection` to `ElementConfigurationContainer`.
