+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/ElementConfigurationContainer.test.tsx
@@ -17,7 +17,7 @@
 import React from 'react';
 import { fireEvent, render, screen } from 'wrappedTestingLibrary';
 
-import ElementConfigurationSection from './ElementConfigurationSection';
+import ElementConfigurationContainer from './ElementConfigurationContainer';
 
 describe('ElementConfigurationSection', () => {
   it('should render elements passed as children', () => {
