+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/SortConfiguration.tsx
@@ -16,10 +16,10 @@
  */
 import * as React from 'react';
 import { FieldArray, useFormikContext } from 'formik';
-
 import Sort from 'views/components/aggregationwizard/elementConfiguration/Sort';
 import ElementConfigurationContainer
   from 'views/components/aggregationwizard/elementConfiguration/ElementConfigurationContainer';
+
 import { WidgetConfigFormValues } from 'views/components/aggregationwizard/WidgetConfigForm';
 
 const SortConfiguration = () => {
